+++ README.md
@@ -28,7 +28,7 @@
 
 ## 공통
 
-- [ ] 영화 데이터를 받아오기 전에 스켈레톤 UI를 먼저 띄운다.
+- [x] 영화 데이터를 받아오기 전에 스켈레톤 UI를 먼저 띄운다.
 
 ## API 명세
