+++ src/view/movieBanner.ts
